Implements Transaction via Composite API
As we know, salesforce can fail at any stage knowadays due a mutitude of
external factors  that is  really hard  to map.  It's a  commom scenario
where we have to do big changes/operations on salesforce records or data
and something unexpectedly fails, then we are screwed as there is no way
back.

Using  the  Awesome  Doctrine  DBAL  api with  the  Composite  API  from
salesforce we managed to have  transactions just like you database would
do it.

                               Brought for you with love by codelicia <3

                                                                    xoxo

Co-authored-by: Alexandre Eher <alexandre.eher@werkspot.nl>
Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>