Solve dependency issue with doctrine/dbal api
On php8 versions, it was allowed to use doctrine/dbal:^3, which changed
a lot of its internal API -- and we do depende on that.

So as a quickfix, we are looking the dbal version to 2, then a proper
fix and version for the newest version will be released soon.

We've also updated the orm version on the way, but the plan is to get
rid of it.

Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>
Co-authored-by: Alexandre Eher <alexandre.eher@werkspot.nl>
Co-authored-by: Shreya <shreya.harkangi@instapro.com>