chore: update commands to Run Infection in CI Workflow
Updated the 'Run Infection' job execution commands in the
GitHub CI workflow (ci.yml) to include a separate instruction
for PHPUnit to generate coverage reports that Infection can make
use of. This change was made to ensure mutation tests are run
with the help of generated coverage reports, thus optimizing the
time it takes to execute the entire CI workflow.

Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>