chore: add assertions to check connection methods result
The tests for the methods `insert`, `update`, and `delete` in `ConnectionWrapperTest.php` were updated to include assertions that validate the returned result. Previous version did not validate if the operation succeeded or not. This change ensures that we're not only executing these operations but also receiving expected outcomes, which helps with ensuring test robustness.

Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>