refactor: Remove unnecessary types for BoundValuesSeparator
The necessity of specifying the types for each parameter in
the 'separateBoundValues' static function has been removed.

As the types are not needed for value separation, it simplifies
both the function and its use in the 'SoqlStatement' file where
the second argument is removed. The work is completed with
additional tests for BoundValuesSeparator ensuring the function
works correctly for different use cases. Therefore, overall code
readability and maintainability are now improved.

Signed-off-by: Jefersson Nathan <malukenho.dev@gmail.com>